requirements.txt
Django==2.2.10
djangorestframework==3.10.2
pytz==2020.1
sqlparse==0.3.1
